Update README.md
Agriculture is a way of life. 
Every living things require food for its survival, and the food has to come from a source. 
In the present scenario, everyone is focused on improving their life standards and has no time in taking care of plants. The motive behind this project is to implement an automated plant nourishing system, which can cover almost all demands of the corresponding plants with only very low human intervention.